requirements.txt
pytest-playwright==0.7.2
openai==2.31.0
dotenv==0.9.9
langchain-core==1.2.28
langchain-openai==1.1.12
playwright-stealth==2.0.3